jupyterlab/hub/requirements.txt
# jupyterhub==5.3.0
jupyterhub-nativeauthenticator==1.3.0
jupyterhub-idle-culler==1.4.0
dockerspawner
jupyter-resource-usage

redis>=4.6.0
psycopg2-binary
python-dotenv

pyjwt
prometheus-client
requests
jinja2
tornado
traitlets
sqlalchemy
pycurl
aiohttp


